p4i integration Integrate change 71509 from 3.0 to main. Description for change 71509:
: Change 71509 by vohi@vohi-hokkaido-nt on 2002/08/21 16:27:09
: 
: 	Explicitely cast handle to long, so that QString::number() can be used correctly on IA64 systems.
: 
: Affected files ...
: 
: ... //depot/qt/3.0/src/kernel/qfont_win.cpp#18 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 71510]

+++ src/kernel/qfont_win.cpp
@@ -333,7 +333,7 @@ void QFontPrivate::load()
     QFontStruct *qfs = 0;
     QString k = key();
     if ( paintdevice )
-	k += QString::number( (Q_LONG)paintdevice->handle() );
+	k += QString::number( (long)(Q_LONG)paintdevice->handle() );
     qfs = fontCache->find( k );
 
     if ( !qfs ) {			// font was never loaded
